refactor: split MCP dashboard tools
